Conflicts:
	StortSpelprojekt/StortSpelprojekt/JsonStructs.h
	StortSpelprojekt/StortSpelprojekt/LevelEdit.cpp
	StortSpelprojekt/StortSpelprojekt/StortSpelprojekt.vcxproj

+++ StortSpelprojekt/StortSpelprojekt/Unit.cpp
@@ -422,7 +422,7 @@ void Unit::CalculatePath()
 
 int Unit::GetApproxDistance(AI::Vec2D target) const
 {
-	return _aStar->GetHeuristicDistance(_tilePosition, target);
+	return static_cast<int>(_aStar->GetHeuristicDistance(_tilePosition, target));
 }
 
 Unit::Unit()
